Fixed from initial run on circle ci

+++ index.js
@@ -29,17 +29,17 @@ async function check(config) {
   const localTags     = config.local.git.tags;
 
   let repoInfo = {}
-  if(localBranches.indexOf(env.CIRCLE_BRANCH) > -1) {
-    repoInfo.branch = env.CIRCLE_BRANCH;
-  }
-  else if(config.local.git.tags && env.CIRCLE_TAG) {
+  if(config.local.git.tags && env.CIRCLE_TAG) {
     repoInfo.tag    = env.CIRCLE_TAG;
   }
-  else if(env.TRAVIS_BRANCH) {
-    repoInfo.branch = env.TRAVIS_BRANCH;
+  else if(localBranches.indexOf(env.CIRCLE_BRANCH) > -1) {
+    repoInfo.branch = env.CIRCLE_BRANCH;
   }
   else if(env.TRAVIS_TAG) {
     repoInfo.tag    = env.TRAVIS_TAG;
+  }
+  else if(env.TRAVIS_BRANCH) {
+    repoInfo.branch = env.TRAVIS_BRANCH;
   }
   else {
     debug("Not in CI env or branch or tag envs missing.");
@@ -111,13 +111,19 @@ async function push(config) {
     })
     const index = await repo.refreshIndex();
 
-    let buildPath = path.join(__dirname, config.local.path);
+    let buildPath = path.relative(process.cwd(), config.local.path);
     const basePath = path.join(repo.workdir(), "builds");
 
     var files = await glob(buildPath+"/**/*");
 
 
     for(fileName of files) {
+      const stats = await fse.lstat(fileName);
+      if(stats.isDirectory()) {
+        debug("Skipping directory ", fileName);
+        continue;
+      }
+
       debug("Writing file to directory: '"+fileName+"'");
       var fileData = await fse.readFile(fileName);
       var relFileName = path.relative(buildPath, fileName);
@@ -140,24 +146,30 @@ async function push(config) {
       return false;
     }
 
-    const latestTag = semverLatest(await fse.readdir(basePath));
-    const latestPath = path.join(basePath, "latest");
-    const latestTagPath = path.join(basePath, latestTag);
+    dirfiles = await fse.readdir(basePath);
+    const latestTag = semverLatest(dirfiles);
 
-    // Remove the old symlink. This might fail if it's the first time this
-    // operation is run.
-    try {
-      await fse.unlink(latestPath);
+    debug("latest tag", latestTag)
+    if(latestTag) {
+      const latestPath = path.join(basePath, "latest");
+      const latestTagPath = path.join(basePath, latestTag);
+
+      // Remove the old symlink. This might fail if it's the first time this
+      // operation is run.
+      try {
+        awa it fse.unlink(latestPath);
+      }
+      catch (err) {
+        // Ignore...
+      }
+
+      var relPath = path.relative(path.dirname(latestPath), latestTagPath)
+      await fse.symlink(relPath, latestPath);
+
+
+      await index.addByPath(path.relative(repo.workdir(), latestPath));
     }
-    catch (err) {
-      // Ignore...
-    }
 
-    var relPath = path.relative(path.dirname(latestPath), latestTagPath)
-    await fse.symlink(relPath, latestPath);
-
-
-    await index.addByPath(path.relative(repo.workdir(), latestPath));
     await index.write();
 
 
@@ -192,6 +204,7 @@ async function push(config) {
     tmpFd.cleanup();
     throw err;
   }
+  tmpFd.cleanup();
 }
 
 module.exports = {
